+++ Sources/CheICalMCP/EventKit/EventKitManager.swift
@@ -1,3 +1,4 @@
+import CheMCPKit
 import CoreLocation
 import EventKit
 import Foundation