+++ Sources/CheICalMCP/EventKit/UndoManager.swift
@@ -1,3 +1,4 @@
+import CheMCPKit
 import EventKit
 import Foundation